+++ extensions/iceworks-lowcode-page-builder/src/components/Material/components/scaffold.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Icon } from '@alifd/next';
-import { PLACEHOLDER_IMG } from '../config/material';
-import { IMaterialScaffold } from '../types/material';
+import { PLACEHOLDER_IMG } from '../config';
+import { IMaterialScaffold } from '../types';
 import * as styles from './scaffold.module.scss';
 
 export const MaterialScaffold: React.FC<{
